삭제과정에서 확인기능 추가, html분기설정 및 view단 로직

+++ posts/views.py
@@ -61,5 +61,8 @@ def post_update(request, post_id):
 def post_delete(request, post_id):
 
     post = Post.objects.get(id=post_id)
-    post.delete()
-    redirect('post-list')
+    if request.method == 'POST':
+        post.delete()
+        return redirect('post-list')
+    else:
+        return render(request, 'posts/post_delete_confirm.html', {'post': post})
